Moved RunTranscoderTest
